// src/app/modules/fianzas/pages/core/fnz-recoleccion-fisicos/interfaces/file-response.interface.ts
export interface FileType {
    aseguradora:    string;
    titular:        string;
    documento:      string;
    nombreOriginal: string;
    nombreCorto:    string;
    version:        number;
    url:            string;
    usuarioAlta:    string;
    eliminado:      boolean;
    _id:            string;
    fechaHoraAlta:  Date;
}