// src/app/shared/helpers/enums/core/documento.enum.ts
export enum EDocumento {
    ACTA_CONSTITUTIVA = 'ACTA_CONSTITUTIVA',
    ACUSE_ENVIO = 'ACUSE_ENVIO',
    ANEXO = 'ANEXO_FIC',
    COMPROBANTE_DOMICILIO_CLIENTE = 'COMPROBANTE_DOMICILIO_CLIENTE',
    COMPROBANTE_DOMICILIO_PROPIETARIO_REAL = 'COMPROBANTE_DOMICILIO_PROPIETARIO_REAL',
    CONFIRMACION_ENTREGA = 'CONFIRMACION_ENTREGA',
    CONSTANCIA_FIRMA_ELECTRONICA = 'CONSTANCIA_FIRMA_ELECTRONICA',
    CONSTANCIA_SITUACION_FISCAL = 'CONSTANCIA_SITUACION_FISCAL',
    CURP = 'CURP',
    ESTADO_CUENTA = 'ESTADO_CUENTA',
    FIC = 'FIC',
    IDENTIFICACION_APODERADO_LEGAL = 'IDENTIFICACION_APODERADO_LEGAL',
    IDENTIFICACION_PROPIETARIO_REAL = 'IDENTIFICACION_PROPIETARIO_REAL',
    PODER_NOTARIAL = 'PODER_NOTARIAL',
    VALIDACION_DE_ORIGINALES = 'VALIDACION_DE_ORIGINALES'
}